shop/requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
astroid==2.12.13
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dateparser==1.1.8
defusedxml==0.7.1
dill==0.3.6
Django==3.2.11
django-admin-decorators==0.1
django-admin-interface==0.25.0
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-filter==22.1
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-render-partial==0.4
django-social-share==2.3.0
djangorestframework==3.12.2
docopt==0.6.2
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
idna==3.4
isort==5.11.4
jet-bridge-base==1.7.17
jet-django==1.7.17
jsonfield==3.1.0
kavenegar==1.1.2
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
oauthlib==3.2.2
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.9
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
Rx==1.6.3
six==1.16.0
social-auth-core==4.4.2
SQLAlchemy==1.4.48
sqlparse==0.4.3
text-unidecode==1.3
tomlkit==0.11.6
tornado==6.3.1
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.13
wcwidth==0.2.6
wrapt==1.14.1
